% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_catch_all.R
\name{all_in_one_time_series}
\alias{all_in_one_time_series}
\title{God version of prediction}
\usage{
all_in_one_time_series(
  original,
  freq = "month",
  steps = 3,
  dlmPoly = 2,
  dlmSeas = 12,
  num.cores = 2,
  error = "mape",
  xreg = NULL,
  a.a.args = list(NULL),
  ets.args = list(NULL),
  tbats.args = list(NULL),
  n.n.args = list(NULL)
)
}
\arguments{
\item{original}{list of time series you want to forecast}

\item{freq}{character string indicating "month" for monthly etc.}

\item{steps}{integer that explains how far out you want forecast(needed for DLM)}

\item{dlmPoly}{integer for the order of polynomial you want in dlm, default = 2}

\item{dlmSeas}{integer for seasonal effect in dlm, default is 12 for monthly}

\item{num.cores}{integer for the number of cores you want to use}

\item{error}{error method you want to use for backtesting}

\item{xreg}{list of matrix/array of exogenous regressors you want to use to forecast only applicable to certain models}

\item{a.a.args}{list of arguments you want to change}

\item{ets.args}{list of arguments you want to change}

\item{tbats.args}{list of arguments you want to change}

\item{n.n.args}{list of arguments you want to change}
}
\value{
a data frame with your series orignal values plus the forecasts with added steps
}
\description{
God version of prediction
}
\examples{
\dontrun{
all_in_one_time_series(mtcars)
}
}
