% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_forecast_by_MAPE.R
\name{best_forecast_by_MAPE}
\alias{best_forecast_by_MAPE}
\title{Finding the best forecast by these error metrics and outputting them}
\usage{
best_forecast_by_MAPE(DF, data, f, DFError, trainStart, ts_data, names)
}
\arguments{
\item{DF}{list of data from time_series_catch}

\item{data}{original data with Date in first column}

\item{f}{integer}

\item{DFError}{list of data frames that contain errors}

\item{trainStart}{numeric vector}

\item{ts_data}{xts matrix}

\item{names}{character vector}
}
\value{
list of data frames with best forecast by that error method
}
\description{
Finding the best forecast by these error metrics and outputting them
}
\examples{
DfMAPE <- best_forecast_by_MAPE(DF, data, f, DFError, trainStart, ts_data, vended_labor)
DF is a list of data frames, data is the original data frame that got iterated through, f is the frequency of the data, DFError is the list of data frames
with the error values in them, trainStart is a vector of the training start date e.g. c(2016, 1), ts_data is the xts matrix of data in time series format,
names is the original data frame with the column names to match.
}
