% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_forecast_by_MSE.R
\name{best_forecast_by_MSE}
\alias{best_forecast_by_MSE}
\title{Finding the best forecast by these error metrics and outputting them}
\usage{
best_forecast_by_MSE(DF, data, f, DFError, trainStart)
}
\arguments{
\item{DF}{list of data from time_series_catch}

\item{data}{original data with Date in first column}

\item{f}{frequency of data i.e. month/day/year 12/365/1, interger}

\item{DFError}{list of data frames that contain errors from error_calculations}

\item{trainStart}{numeric vector, i.e. c(2016, 1) for Jan 2016}
}
\value{
list of data frames with best forecast by that error method
}
\description{
Finding the best forecast by these error metrics and outputting them
}
\examples{
DfMSE <- best_forecast_by_MSE(DF, data, f, DFError, trainStart)

}
