% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_forecast_by_RMSE.R
\name{best_forecast_by_RMSE}
\alias{best_forecast_by_RMSE}
\title{Finding the best forecast by these error metrics and outputting them (3)}
\usage{
best_forecast_by_RMSE(DF, data, f, DFError, trainStart)
}
\arguments{
\item{DF}{list of data from time_series_catch}

\item{data}{original data with Date in first column}

\item{f}{frequency of data i.e. month/day/year 12/365/1, interger}

\item{DFError}{list of data frames that contain errors from error_calculations}

\item{trainStart}{numeric vector, i.e. c(2016, 1) for Jan 2016}
}
\value{
list of data frames with best forecast by that error method
}
\description{
Finding the best forecast by these error metrics and outputting them (3)
}
\examples{
DfRMSE <- best_forecast_by_RMSE(DF, data, f, DFError, trainStart, ts_data, vended_labor)
DF is a list of data frames, data is the original data frame that got iterated through, f is the frequency of the data, DFError is the list of data frames
with the error values in them, trainStart is a vector of the training start date e.g. c(2016, 1), ts_data is the xts matrix of data in time series format,
names is the original data frame with the column names to match.
}
