% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous_regressors.R
\name{exogenous_regressors}
\alias{exogenous_regressors}
\title{building our exogenous regressors for our arimax and neural net models term, moq, quarter}
\usage{
exogenous_regressors(data)
}
\arguments{
\item{ts_data}{xts matrix with variable Date in first column}
}
\value{
a matrix of xregs based on h
}
\description{
building our exogenous regressors for our arimax and neural net models term, moq, quarter
}
\examples{
xreg <- exogenous_regressors(ts_data) makes term variables, quarter, and MoQ variables
}
