% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_best_error.R
\name{extract_best_error}
\alias{extract_best_error}
\title{getting error values from the best model}
\usage{
extract_best_error(data, DFError, x = 1)
}
\arguments{
\item{data}{original data}

\item{DFError}{list of data frames that contain errors}

\item{x}{indicator for mape(default)=1, x=2 rmse, x=3 mse}
}
\value{
numeric vector of best error values corresponding to n-1 column
}
\description{
getting error values from the best model
}
\examples{
q <- extract_best_error(data, DFError)
}
