% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_time_series_catch_all.R
\name{fast_time_series_catch_all}
\alias{fast_time_series_catch_all}
\title{ONLY USE IF YOU ALREADY RAN THE FIRST MODEL(time_series_catch_all) AND HAVE A VEcTOR OF MODEL NAMES}
\usage{
fast_time_series_catch_all(data, f, h, trainStart, trainEnd, test, n, cols,
  bestModels)
}
\arguments{
\item{data}{original data containing Date in first column}

\item{f}{integer}

\item{h}{integer}

\item{trainStart}{numeric vector}

\item{trainEnd}{numeric vector}

\item{test}{numeric vector}

\item{n}{integer}

\item{cols}{character vector}

\item{bestModels}{vector of best modelNames equal to columns-1}
}
\value{
a list of data frames that only contain the best model found by using the best_model_name function
}
\description{
ONLY USE IF YOU ALREADY RAN THE FIRST MODEL(time_series_catch_all) AND HAVE A VEcTOR OF MODEL NAMES
}
\examples{
the fast version of time-series catch all, can oly be run after you run the first model due to the dependency on the bestModels variable
as it will only run those types of models in the future. To use this you need to take names_of_best_model and do as.character(unlist(modelList)) and
input that as your bestModels variable for this function.
}
