% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_catch_all.R
\name{forecast_function}
\alias{forecast_function}
\title{forecast function that forecasts out how many steps you look forward}
\usage{
forecast_function(
  orig_x,
  models,
  freq = freq,
  steps = steps,
  dlmPoly = dlmPoly,
  dlmSeas = dlmSeas,
  num.cores = num.cores,
  error = error,
  xreg = xreg,
  a.a.args = a.a.args,
  ets.args = ets.args,
  tbats.args = tbats.args,
  n.n.args = n.n.args
)
}
\arguments{
\item{orig_x}{list of time series you want to forecast}

\item{models}{list of models}

\item{freq}{character string indicating "month" for monthly etc.}

\item{steps}{integer that explains how far out you want forecast(needed for DLM)}

\item{dlmPoly}{integer for the order of polynomial you want in dlm, default = 2}

\item{dlmSeas}{integer for seasonal effect in dlm, default is 12 for monthly}

\item{num.cores}{integer for the number of cores you want to use}

\item{error}{either is "smape" or "mape", default is "mape"}

\item{xreg}{list of matrix/array of exogenous regressors you want to use to forecast only applicable to certain models}

\item{a.a.args}{list of arguments you want to change}

\item{ets.args}{list of arguments you want to change}

\item{tbats.args}{list of arguments you want to change}

\item{n.n.args}{list of arguments you want to change}
}
\value{
list of models for models
}
\description{
forecast function that forecasts out how many steps you look forward
}
\examples{
\dontrun{
test <- replicate(5, list(ldeaths))
test_model <- par_time_series_catch(test, num.cores = 12)
}

}
