% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_catch_all.R
\name{mean_prediction}
\alias{mean_prediction}
\title{mean prediction for data that isnt very good}
\usage{
mean_prediction(df, steps = 3)
}
\arguments{
\item{df}{the dataframe you want to predict with the mean}

\item{steps}{the number of steps you are looking forward}
}
\value{
a data frame of your predictions
}
\description{
mean prediction for data that isnt very good
}
\examples{
\dontrun{
test_model <- mean_prediction(mtcars, steps=3)
}
}
