% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_build.R
\name{model.build.12}
\alias{model.build.12}
\title{Model building for DLM}
\usage{
model.build.12(p)
}
\arguments{
\item{p}{Used in DLM modelling only}
}
\value{
DLM
}
\description{
Model building for DLM
}
\examples{
Monthly seasonal model
}
