% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_build_3.R
\name{model.build.3}
\alias{model.build.3}
\title{Model building for DLM}
\usage{
model.build.3(p)
}
\arguments{
\item{p}{used in DLM modelling}
}
\value{
DLM
}
\description{
Model building for DLM
}
\examples{
Quarterly seasonal model
}
