% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_build.R
\name{model.build}
\alias{model.build}
\title{Model building for DLM}
\usage{
model.build(p)
}
\arguments{
\item{p}{Used in DLM modelling only}
}
\value{
DLM
}
\description{
Model building for DLM
}
\examples{
makes a quarterly seasonal model in the dlm sense with 2 polynomial terms
}
