% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_catch_all.R
\name{par_time_series_catch}
\alias{par_time_series_catch}
\title{The big dawg that will forecast with the best of them}
\usage{
par_time_series_catch(x, freq = "month", steps = 3, dlmPoly = 2,
  dlmSeas = 12, num.cores = 2, a.a.args = list(NULL),
  ets.args = list(NULL), tbats.args = list(NULL),
  prophet.args = list(NULL), n.n.args = list(NULL))
}
\arguments{
\item{x}{list of time series you want to forecast}

\item{freq}{character string indicating "month" for monthly etc.}

\item{steps}{integer that explains how far out you want forecast(needed for DLM)}

\item{dlmPoly}{integer for the order of polynomial you want in dlm, default = 2}

\item{dlmSeas}{integer for seasonal effect in dlm, default is 12 for monthly}

\item{num.cores}{integer for the number of cores you want to use}

\item{a.a.args}{list of arguments you want to change}

\item{ets.args}{list of arguments you want to change}

\item{tbats.args}{list of arguments you want to change}

\item{prophet.args}{list of arguments you want to change}

\item{n.n.args}{list of arguments you want to change}
}
\value{
list of models for eacg ts
}
\description{
The big dawg that will forecast with the best of them
}
\examples{
hundo <- replicate(5, list(ldeaths))
test <- par_time_series_catch(hundo, num.cores = 12)
cl <- makeCluster(getOption("cl.cores", 12))
test_fit <- parLapply(cl, test, extract_model_fit_forecast)
stopCluster(cl)
}
