% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dlm_model.R
\name{ts_dlm_model}
\alias{ts_dlm_model}
\title{dlm model}
\usage{
ts_dlm_model(time, n_train, h)
}
\arguments{
\item{time}{ts}

\item{n_train}{ts}

\item{h}{integer from 1-3}
}
\value{
forecasted values based on h
}
\description{
dlm model
}
\examples{
gets called in the time_series_catch_all function to run DLM models.
}
