% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_prohpet.R
\name{ts_prophet}
\alias{ts_prophet}
\title{prohpet model}
\usage{
ts_prophet(data, h, i)
}
\arguments{
\item{data}{original data with Date in first column}

\item{h}{integer from 1-3}

\item{i}{iteration for loop}
}
\value{
forecasted values from prophet
}
\description{
prohpet model
}
\examples{
Uses the data, h and the i value from the loop in time_series_catch_all to get the forecast values.
}
