% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_catch_all.R
\name{values_replaced}
\alias{values_replaced}
\title{tells you what values got changed in your dataframe}
\usage{
values_replaced(new, old)
}
\arguments{
\item{new}{dataframe that has imputed values}

\item{old}{dataframe of same dim with original values}
}
\value{
prints whats changed from new to old
}
\description{
tells you what values got changed in your dataframe
}
\examples{
\dontrun{
values_replaced(mtcars, mtcars*.2s)
}
}
