% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_forecast_by_MAPE.R
\name{best_forecast_by_MAPE}
\alias{best_forecast_by_MAPE}
\title{Finding the best forecast by MAPE and outputting them into a list of data frames}
\usage{
best_forecast_by_MAPE(DF, data, f, DFError, trainStart)
}
\arguments{
\item{DF}{list of data frames from time_series_catch function}

\item{data}{original data with Date in first column}

\item{f}{frequency of data i.e. month/day/year 12/365/1, interger}

\item{DFError}{list of data frames that contain errors from error_calculations}

\item{trainStart}{numeric vector, i.e. c(2016, 1) for Jan 2016}
}
\value{
list of data frames with best forecast by that error method
}
\description{
Finding the best forecast by MAPE and outputting them into a list of data frames
}
\examples{
DfMAPE <- best_forecast_by_MAPE(DF, data, f, DFError, trainStart)

}
