% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_catch_all.R
\name{extract_model_fit_forecast}
\alias{extract_model_fit_forecast}
\title{Extract the numeric array of fitted values with the forecast added on to the end.}
\usage{
extract_model_fit_forecast(x, steps = 3, xreg = NULL, freq = "month",
  num.cores = 2)
}
\arguments{
\item{x}{list of models from par_catch_all}

\item{steps}{integer of how many steps out you want to forecast}

\item{xreg}{list of matrix/array of exogenous regressors you want to use to forecast only applicable to certain models}

\item{freq}{string of characters indicating e.g. "month" for monthly etc}
}
\value{
list of numeric arrays with forecasts appended
}
\description{
Extract the numeric array of fitted values with the forecast added on to the end.
}
\examples{
 hundo <- replicate(5, list(ldeaths))
test <- par_time_series_catch(hundo, num.cores = 12)
cl <- makeCluster(getOption("cl.cores", 12))
test_fit <- parLapply(cl, test, extract_model_fit_forecast)
stopCluster(cl)
}
