% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_time_series_catch_all.R
\name{fast_time_series_catch_all}
\alias{fast_time_series_catch_all}
\title{ONLY USE IF YOU ALREADY RAN THE FIRST MODEL(time_series_catch_all) AND HAVE A VEcTOR OF MODEL NAMES}
\usage{
fast_time_series_catch_all(data, f, h, trainStart, trainEnd, test, n,
  bestModels, is.VL = FALSE, seasonaility = "Monthly")
}
\arguments{
\item{data}{original data with Date in first column}

\item{f}{frequency of data i.e. month/day/year 12/365/1, interger}

\item{h}{forecasting steps up to 3, integer}

\item{trainStart}{numeric vector, i.e. c(2016, 1) for Jan 2016}

\item{trainEnd}{numeric vector, i.e. c(2016, 1) for Jan 2016}

\item{test}{numeric vector, i.e. c(2016, 1) for Jan 2016}

\item{n}{integer, number of cores you want to use}

\item{bestModels}{vector of best modelNames, comes from best_model_name function}

\item{is.VL}{if the data is vended labor flag it with is.VL = TRUE}

\item{seasonaility}{if quarterly choose seasonaility = "Quarter" otherwise dont change default is monthly}
}
\value{
a list of data frames that only contain the best model found by using the best_model_name function
}
\description{
ONLY USE IF YOU ALREADY RAN THE FIRST MODEL(time_series_catch_all) AND HAVE A VEcTOR OF MODEL NAMES
}
\examples{
the fast version of time-series catch all, can oly be run after you run the first model due to the dependency on the bestModels variable
as it will only run those types of models in the future. To use this you need to take the output from names_of_best_model and input it into the
bestModels variable. therefore it will only run the functions that were chosen to be the best model.
}
