% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_catch.R
\name{time_series_catch}
\alias{time_series_catch}
\title{time series catch all}
\usage{
time_series_catch(data, f, h, trainStart, trainEnd, test, n, cols)
}
\arguments{
\item{data}{original data with Date in first column}

\item{f}{integer}

\item{h}{integer}

\item{trainStart}{numeric vector}

\item{trainEnd}{numeric vector}

\item{test}{numeric vector}

\item{n}{integer}

\item{cols}{character vector}
}
\value{
a list of data frames with all the forecasts in each frame for each respective column
}
\description{
time series catch all
}
\examples{
# Function call for time series catch all with output DF is a list of forecasted fata frames
DF <- time_series_catch(data, f, h, trainStart, trainEnd, test, n, as.vector(unique)) in which data is the orignal data with the variable Date in the first column
f is frequency, h is forecasted steps up to 3, trainStart/trainEnd/test are all date vectors specified with c(2016, 1), n is the number of cores to use for pp
cols is a vector of equal length of the data column minus 1 which holds the names of all the unique column names.
}
